moving actions from ProjectDetailsHelpers to ProjectDetailsActions
