+++ src/js/helpers/GitApi.js
@@ -177,7 +177,7 @@ export default function GitApi(directory) {
  * splits the repo url to get repo name
  */
 export const parseRepoUrl = (ulr) => {
-  const repoName = ulr.trim().match(/^(\w*)(:\/\/|@)([^/:]+)[/:]([^/:]+)\/(.+)(.git)?$/) || [''];
+  const repoName = ulr.trim().match(/^(\w*)(:\/\/|@)([^/:]+)[/:]([^/:]+)\/(.+)(.git)?$/) || ['']; // match even if `.git` extension missing
   const repoInfo = {
     name: repoName[5],
     user: repoName[4],
